cpu_input_boost: [FIX] introduce negative display state aware boosting
Negative boosting can be helpful for battery. To improve idle battery
drain, set the global boost to -15 and the boost for all app processes
and unimportant system processes to -35. When the display is on, remove
the negative boosts with the exception of applying a -15 boost to
background apps. All values may be tuned as desired in the config and
sysfs module parameters.

Note: commit e71c4255169ddacf3472f09b0c7fbe0fe808214e is required for
any of the negative boosts to work. Without it, they will simply be
ignored by schedtune.

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>